backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
twilio==8.10.0
python-dotenv==1.0.0
httpx==0.25.2
websockets==12.0
aiosmtplib==3.0.1
python-dateutil==2.8.2
